+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto_classes/Auto1.java
@@ -9,9 +9,6 @@ import org.firstinspires.ftc.teamcode.team_classes.Robot;
 
 @Disabled
 public class Auto1 extends OpMode {
-    
-
-
     private ElapsedTime runtime = new ElapsedTime();
     private Robot Robot = new Robot();
