+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto_classes/drivers/DirectMecanum.java
@@ -1,16 +1,15 @@
-package org.firstinspires.ftc.teamcode.auto_classes;
+package org.firstinspires.ftc.teamcode.auto_classes.drivers;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.robotcore.external.navigation.Position;
 import org.firstinspires.ftc.teamcode.general_classes.Position2DAngle;
 import org.firstinspires.ftc.teamcode.team_classes.Robot;
 
 
-@TeleOp(name="Driver Luke", group="Driver") //fix this
-public class DriverLuke extends OpMode {
+@TeleOp(name="Default Direct", group="9108") //fix this
+public class DirectMecanum extends OpMode {
 
     private ElapsedTime runtime = new ElapsedTime();
     private Robot Robot = new Robot();
@@ -29,6 +28,7 @@ public class DriverLuke extends OpMode {
     @Override
     public void start() {
         runtime.reset();
+        telemetry.setAutoClear(true);
     }
 
     boolean on;
@@ -36,11 +36,10 @@ public class DriverLuke extends OpMode {
     //Initialized by: After Start, Before Stop / loops
     @Override
     public void loop() {
-        //Joysticks
+        //DRIVING
         double drivey = gamepad1.left_stick_y;
         double drivex = -gamepad1.left_stick_x;
         double turn = gamepad1.right_stick_x;
-        double PERCENTAGE = Math.hypot(drivey, drivex);
         if (Math.abs((double)drivey) < .05) {
             drivey = 0;
         }
@@ -50,21 +49,39 @@ public class DriverLuke extends OpMode {
         if (Math.abs((double)turn) < .05) {
             turn = 0;
         }
-        double Heading = (Math.toDegrees(Robot.IMU.getHeadingRadians()));
-        Position2DAngle relativeValues = Robot.DCGm.relativeValues(new Position2DAngle(drivex,drivey,turn), Heading);
-        Position2DAngle adjustedValues = new Position2DAngle( relativeValues.X * PERCENTAGE
-                                                            , relativeValues.Y * PERCENTAGE
-                                                            , relativeValues.ANGLE);
-        Robot.DCGm.driveToPositionAngle(adjustedValues, true);
+        Robot.DCGm.driveToPositionAngle(new Position2DAngle(drivex,drivey,turn), true);
 
-        //Buttons
-        //DPad_UP
+        /*
+        //Lift
+        double lift = gamepad2.left_stick_y;
+        double speed = lift * 100;
+        Robot.DCGl.raiseToInch(0.1,speed);
 
-        //DPad_RIGHT
+        if (gamepad2.a) {
+            Robot.DCGl.raiseToBlock(-1,75);
+        }
+        if (gamepad2.y) {
+            Robot.DCGl.raiseToBlock(1,75);
+        }
 
-        //DPad_D
+        //Intake
+        if (on == true) {
+            Robot.DCGi.setPower(new double[]{1});
+        }
+        if (on == false) { 
+            Robot.DCGi.setPower(new double[]{0});
+        }
 
+        if (gamepad2.left_trigger == 1) {
+            on = true;
+        }
 
+        if (gamepad2.right_trigger == 1) {
+            on = false;
+        }
+         */
+
+        Robot.DCGm.composeTelemetry(telemetry);
         telemetry.addData("Status", "Run Time: " + runtime.toString());
     }
